gerror: Add generic errors
Libraries and applications tend to abuse Gio's G_IO_ERROR_FAILED error
code, because they don't care about the error domain and code, but
only about the fact that the error happened. So just add a generic
error domain and code, and several functions to accommodate the
developers.

Not sure if this is a good practice, though. And if GLib should have a
thing like a generic error. But the other way is to continue abusing
G_IO_ERROR_FAILED.